requirements.txt
Django==2.1.2
django-markdown-deux==1.0.5
django-uuslug==1.1.8
markdown2==2.3.6
python-slugify==1.2.6
pytz==2018.5
Unidecode==1.0.22
